+++ apps/website/src/routes/docs/headless/tabs/examples/dynamic.tsx
@@ -4,32 +4,55 @@ import styles from '.././index.css?inline';
 
 export default component$(() => {
   useStyles$(styles);
-  const tabsState = useStore(['Dynamic Tab 1', 'Dynamic Tab 2', 'Dynamic Tab 3']);
+  const store = useStore({
+    tabs: ['Tab 1', 'Tab 2', 'Tab 3'],
+  });
   return (
     <>
-      <div class="tabs-example mr-auto">
-        <Tabs.Root>
-          <Tabs.List>
-            {tabsState.map((tab) => (
-              <Tabs.Tab key={tab}>{tab}</Tabs.Tab>
-            ))}
-          </Tabs.List>
-          {tabsState.map((tab) => (
-            <Tabs.Panel key={tab}>{tab} Panel</Tabs.Panel>
-          ))}
-        </Tabs.Root>
-
+      <div class="vertical-tabs-example mr-auto">
         <button
-          class="mt-4 font-bold text-red-600"
+          class={cn('mb-4 w-48', buttonVariants({ look: 'outline', size: 'sm' }))}
           onClick$={() => {
-            if (tabsState.length > 1) {
-              tabsState.splice(0, 1);
-            }
+            store.tabs.unshift(`Added Tab`);
           }}
         >
-          Remove First Tab
+          Add Tab before
+        </button>
+        <br />
+        <Tabs.Root vertical class="flex space-x-6">
+          <Tabs.List class="flex w-fit flex-col">
+            {store.tabs.map((tab, i) => (
+              <Tabs.Tab key={i}>
+                {tab}{' '}
+                <span
+                  onClick$={() => {
+                    if (store.tabs.length > 1) {
+                      store.tabs.splice(i, 1);
+                    }
+                  }}
+                >
+                  x
+                </span>
+              </Tabs.Tab>
+            ))}
+          </Tabs.List>
+
+          {store.tabs.map((tab, i) => {
+            return <Tabs.Panel key={i}>{tab} Panel</Tabs.Panel>;
+          })}
+        </Tabs.Root>
+        <button
+          class={cn('w-48', buttonVariants({ look: 'outline', size: 'sm' }))}
+          onClick$={() => {
+            store.tabs.push(`Added Tab`);
+          }}
+        >
+          Add Tab after
         </button>
       </div>
     </>
   );
 });
+
+import { buttonVariants } from '@qwik-ui/styled';
+import { cn } from '@qwik-ui/utils';