+++ packages/kit-headless/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 0.6.0
+
+### Minor Changes
+
+- ✨ toggle and togglegroup headless and styled components (by [@steffanek](https://github.com/steffanek) in [#957](https://github.com/qwikifiers/qwik-ui/pull/957))
+
+- ✨ carousel reaches beta state (by [@thejackshelton](https://github.com/thejackshelton) in [#965](https://github.com/qwikifiers/qwik-ui/pull/965))
+
+  feat: stepper component added as a configuration for the carousel
+
+  feat: vertical carousels are now supported
+
+  feat: progress bar gets a major refactor and is backwards compatible
+
+  docs: fixed theme issues and improved prefetching
+
 ## 0.5.2
 
 ### Patch Changes